Move ScoreDetail and ScoreList to own component
