+++ BoardGameScorer.WebUI.Server/Components/Pages/KingDomino/KingDomino.razor
@@ -6,34 +6,22 @@
 <EditForm EditContext="_editContext" FormName="myForm" OnValidSubmit="HandleSubmit">
 	<DataAnnotationsValidator />
 	<ValidationSummary />
-	<div class="form-group mb-3">
+	<div class="mb-3">
+		<label for="terrain" class="form-label">--Select Terrain--</label>
 		<InputSelect id="terrain" class="form-select" @bind-Value=Model!.Terrain>
-			@if(Model!.Terrain is null)
-			{
-				<option selected value="">--Select Terrain--</option>
-			}
-			else
-			{
-				<option value=""></option>
-			}
-			@foreach (var terrain in Terrains)
-			{
-				if (terrain == "Castle")
-				{
-					<option hidden>--Select Terrain--</option>
-				}
-				else
+			<optgroup label="--Select Terrain--">
+				@foreach (var terrain in Enum.GetValues(typeof(Terrain)))
 				{
 					<option value="@terrain">@terrain</option>
 				}
-			}
+			</optgroup>
 		</InputSelect>
-		<ValidationMessage For="@(() => @Model.Terrain)" />
+		<ValidationMessage For="@(() =>  @Model!.Terrain)" />
 	</div>
 
 	<div class="form-group mb-3">
 		<label class="form-label" for="tiles">How many tiles?</label>
-		<NumberInput TValue="int" id="tiles" @bind-Value=Model.Tiles EnableMinMax="true" Min="1" Max="24" Placeholder="Enter number of tiles"></NumberInput>
+		<NumberInput TValue="int" id="tiles" @bind-Value=Model.Tiles EnableMinMax=true Min="1" Max="24" required></NumberInput>
 		<ValidationMessage For="@(() => Model.Tiles)" />
 	</div>
 
@@ -41,7 +29,7 @@
 		<label class="form-label" for="crown">How many crowns?</label>
 		<InputSelect id="crown" class="form-select" @bind-Value=Model.Crowns>
 
-			@foreach (var crown in Crowns)
+			@foreach (var crown in (int[])Enum.GetValues(typeof(TileWithCrown)))
 			{
 				<option value="@crown">@crown</option>
 			}
@@ -49,35 +37,9 @@
 		</InputSelect>
 		<ValidationMessage For="@(() => Model!.Crowns)" />
 	</div>
-
 	<Button Type="ButtonType.Submit" Color="ButtonColor.Light">
 		Confirm
 	</Button>
 </EditForm>
 
-
-@if (registration.Count > 0)
-{
-	<table class="table">
-		<thead>
-			<tr>
-				<th scope="col"></th>
-				<th scope="col">Terrain</th>
-				<th scope="col">No of tiles</th>
-				<th scope="col">No of crowns</th>
-			</tr>
-		</thead>
-		<tbody>
-
-			@foreach (var game in registration)
-			{
-				<tr>
-					<th scope="row"></th>
-					<td>@game.Terrain</td>
-					<td>@game.Tiles</td>
-					<td>@game.Crowns</td>
-				</tr>
-			}
-		</tbody>
-	</table>
-}
+<KingDominoScoreList Registration="@_registration" />
