+++ BoardGameScorer.WebUI.Server/Components/_Imports.razor
@@ -20,3 +20,7 @@
 @using BoardGameScorer.Infrastructure.Data
 @using BoardGameScorer.WebUI.Server
 @using BoardGameScorer.WebUI.Server.Components
+@using BoardGameScorer.WebUI.Server.Components.Account
+@using BoardGameScorer.WebUI.Server.Components.Layout
+@using BoardGameScorer.WebUI.Server.Components.Pages
+@using BoardGameScorer.WebUI.Server.Components.Pages.KingDomino
